+++ b2sdk/raw_api.py
@@ -20,9 +20,6 @@ from .encryption.setting import EncryptionMode, EncryptionSetting
 from .file_lock import BucketRetentionSetting, FileRetentionSetting, LegalHold
 from .utils import b2_url_encode
 from b2sdk.http_constants import FILE_INFO_HEADER_PREFIX
-from b2sdk.http_constants import SRC_LAST_MODIFIED_MILLIS
-assert SRC_LAST_MODIFIED_MILLIS  # workaround for pyflakes: disable
-# this is here because CLI depends on it TODO: remove after fixing imports in CLI
 
 # All supported realms
 REALM_URLS = {