+++ CHANGELOG.md
@@ -8,9 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * The `importlib-metadata` requirement is less strictly bound now (just >=3.3.0 for python > 3.5).
+* `B2Api` `update_file_legal_hold` and `update_file_retention_setting` now return the set values 
 
 ### Added
 * `BucketIdNotFound` thrown based on B2 cloud response
+* `_clone` method to `FileVersion` and `DownloadVersion`
+* `delete`, `update_legal_hold`, `update_retention` and `download` methods added to `FileVersion`
+
+### Fixed
+* FileSimulator returns special file info headers properly
 
 ## [1.11.0] - 2021-06-24