+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * FileSimulator returns special file info headers properly
 
+### Removed
+* One unused import.
+
 ## [1.11.0] - 2021-06-24
 
 ### Changed